ACMS-3417: Acquia cms telemetry changes as per acquia connector.
